reduced code duplicity
